Starting provider via commands::run::execute
