+++ sfw-provider/src/built_info.rs
@@ -1,0 +1,2 @@
+// The file has been placed there by the build script.
+include!(concat!(env!("OUT_DIR"), "/built.rs"));
