+++ sfw-provider/src/commands/mod.rs
@@ -1,0 +1,5 @@
+use crate::config::Config;
+use clap::ArgMatches;
+
+pub mod init;
+pub mod run;
